[Connectors Admin] Separate cli and admin lib
Required for https://github.com/dust-tt/tasks/issues/517

Uses INTERACTIVE_CLI env var to distinguish CLI from API
